+++ tests/test_dist_enum.py
@@ -247,3 +247,87 @@ def test_burst_ids_consistent_between_pre_and_post(mgrs_tile_ids: list[str], tra
         df_pre = df_product[df_product['input_category'] == 'pre'].reset_index(drop=True)
         df_post = df_product[df_product['input_category'] == 'post'].reset_index(drop=True)
         assert sorted(df_pre['jpl_burst_id'].unique().tolist()) == sorted(df_post['jpl_burst_id'].unique().tolist())
+
+
+@pytest.mark.integration
+def test_dist_enum_one_with_multi_window_with_multiple_polarizations_and_asf_daac() -> None:
+    """Test enumeration of 1 product with multi_window strategy with multiple dual polarization data.
+
+    Context: MGRS Tile 20TLP: https://search.asf.alaska.edu/#/?polygon=
+    POLYGON((-65.5041%2044.226,-65.4632%2043.2383,-64.1113%2043.2594,-64.1298%2044.2478,-65.5041%2044.226))
+    &start=2025-09-18T07:00:00Z&end=2025-09-20T06:59:59Z&resultsLoaded=true&zoom=8.078
+    &center=-63.112,42.844&dataset=OPERA-S1&productTypes=RTC
+    &granule=OPERA_L2_RTC-S1_T171-365960-IW2_20250919T102314Z_20250919T135744Z_S1C_30_v1.0
+    """
+    df_product = enumerate_one_dist_s1_product(
+        '20TLP',
+        track_number=171,
+        post_date='2025-09-19',
+        lookback_strategy='multi_window',
+        # Need to look back further for validat VV+VH data
+        delta_lookback_days=(1460, 1095, 730, 365),
+        max_pre_imgs_per_burst=(3, 3, 3, 4),
+    )
+
+    assert sorted(df_product.polarizations.unique().tolist()) == ['HH+HV', 'VV+VH']
+
+    df_sample_vvvh_burst = df_product[df_product.jpl_burst_id == 'T171-365965-IW3'].reset_index(drop=True)
+    dates_for_sample_vvvh_burst = sorted(df_sample_vvvh_burst['acq_date_for_mgrs_pass'].unique().tolist())
+    # Note the last date is the post date
+    expected_dates = ['2020-09-21', '2021-05-19', '2021-05-31', '2025-09-19']
+    assert dates_for_sample_vvvh_burst == expected_dates
+
+    # Check baseline data
+    # The post image is VV+VH
+    # Ref: https://search.asf.alaska.edu/#/?dataset=OPERA-S1&productTypes=RTC&operaBurstID=T171_365965_IW3
+    opera_ids = df_product.opera_id.unique().tolist()
+    opera_ids_trunc = ['_'.join(op_id.split('_')[:4]) for op_id in opera_ids]
+    # another VV+VH image
+    assert 'OPERA_L2_RTC-S1_T171-365965-IW3_20250826T102328Z' in opera_ids_trunc
+    # a HH+HV image
+    assert 'OPERA_L2_RTC-S1_T171-365965-IW3_20240427T102443Z' not in opera_ids_trunc
+
+
+@pytest.mark.integration
+def test_dist_enum_one_with_multi_window_with_asf_daac() -> None:
+    df_product = enumerate_one_dist_s1_product(
+        '11SLT',
+        track_number=144,
+        post_date='2025-06-19',
+        lookback_strategy='multi_window',
+        delta_lookback_days=(1095, 730, 365),
+        max_pre_imgs_per_burst=(3, 3, 4),
+    )
+    burst_ids_expected = [
+        'T144-308024-IW1',
+        'T144-308025-IW1',
+        'T144-308026-IW1',
+        'T144-308027-IW1',
+        'T144-308028-IW1',
+        'T144-308029-IW1',
+        'T144-308030-IW1',
+        'T144-308031-IW1',
+    ]
+    assert sorted(df_product['jpl_burst_id'].unique().tolist()) == sorted(burst_ids_expected)
+
+    post_ind = df_product.input_category == 'post'
+    df_product_post = df_product[post_ind].reset_index(drop=True)
+
+    pre_ind = df_product.input_category == 'pre'
+    df_product_pre = df_product[pre_ind].reset_index(drop=True)
+
+    pre_dates_expected = [
+        '2024-06-12',
+        '2024-05-31',
+        '2024-05-19',
+        '2024-05-07',
+        '2023-06-18',
+        '2023-06-06',
+        '2023-05-25',
+        '2022-06-11',
+        '2022-05-30',
+        '2022-05-18',
+    ]
+    assert sorted(df_product_pre['acq_date_for_mgrs_pass'].unique().tolist()) == sorted(pre_dates_expected)
+
+    assert df_product_post['acq_date_for_mgrs_pass'].unique().tolist() == ['2025-06-19']